Permissions Manager
This implementation uses a library (https://github.com/QuickPermissions/QuickPermissions-Kotlin) that uses the old API to request permissions on Android. The new API for requesting permissions in Android is unfortunately very tied to the view, since callbacks have to be registered before the fragment's onCreate event is called.

I wrapped the QuickPermissions library for use with Coroutines.

I added the Kotlin extensions for Fragments and Activities in case we want to switch to the new API.